Create OOP
This code is an initial version, thus there is literally no optimization. Also, since it is an initial version many options are not available. For countries, not all of them are available, for cities not all of them are available, for places not all of them are available, for transportation not all of them are available too.